requirements.txt
Django==3.1
django-rest-swagger==2.2.0
djangorestframework==3.12
django-allauth
django-rest-auth==0.9.5
drf-yasg==1.21.3  
pillow==9.2
gunicorn==20.0.4
dj-rest-auth==2.2.5 